backend/requirements.txt
fastapi
uvicorn
pandas
aiohttp
duckdb
python-dotenv
python-multipart
openpyxl

